fix: Remove usage of non-existent agent/card method (#64)
This PR replaces occurrences of `agent/card` method invocation which
does not exist in the specification with
`agent/getAuthenticatedExtendedCard` that does. Furthermore for tests in
`tests/mandatory/protocol/test_a2a_v030_new_methods.py` it replaces call
to `sut_client.get_agent_card()` with usage of `agent_card_data`
fixture. The result of this call was used to detect if agent supports
the authenticated card which is incorrect.

Fixes #61